Travis reuses cache among jobs
There's this issue travis-ci/travis-ci#7590 where Travis intentionally
does not reuse the cache for jobs with different environment variables.
A pull request travis-ci/travis-build#1265 is open for a few months so I
decided to go the other route and avoid using environment variables,
which I discovered to be possible when reading several repositories and
finally finding ReactiveCollections, from which I copied and adapted the
travis and build scripts.

Ran `travis setup releases` and signed in with GitHub to generate a new
personal access token and have it encrypted.
Reference: https://github.com/travis-ci/travis-ci/issues/2457