Allows setting HTTP body to the request
- `Task` is now a struct since it better matches the use cases for this.
For now we only have body and parameters as optionals and it already
means we need 2^2 = 4 enum case combinations to match all use cases (no
body and no parameters, only body, only parameters, and body+parameters)
- HTTPMethod is now a property of Task instead of Target. The method to
use is something individually related to each case in a Target, so most
usually users from this framework will be writing switch-cases inside
`var method` - and another switch-case inside `var task`. Thus, now the
only expected switch-case is for `var task`.
- The use case where it could be interesting to have a separate `var
method` is when all endpoints for a target use the same method, which
is probably GET when it happens. So this facility was kept by having GET
as the default method for `Task`s.
- `headers` should also be part of `Task`, but this one is not so
usually *individually* related to each Target case, as most often it is
reused by all cases or with small nuances. On the other hand,
`parameters` also tend to be reused, but only a few of them (like an API
key) so it still makes sense to have them as part of the `Task`.