Fixed SPM
Several fixes here, several readings to fix it because Xcode (and swift
binary) insist to throw unhelpful warnings and errors, but yay we did
it! The syntax in the Package.swift file has changed a lot with recent
swift versions and even `swift package init --type library` doesn't seem
to use the most recent one, so after a lot of Googling and trial and
error, voilà a working version. Specifying the paths for the targets
was a very important step! For tests it isn't needed but then we would
need to create a nested directory containing all test files, inside the
directory already named "Tests"...