Fixed unable to implement ProductionTargetType separately
If a project implements ProductionTargetType in the main target, and
wants to implement TargetType in the tests target, it would fail because
the compiler seems to be stuck with the main target's original
implementation, wanting the TargetStub to be a ProductionStub instead of
a Stub.

Unfortunately, this was fixed by replacing a compile-time constraint by
a runtime error.